+++ JiraAgilePS/Public/Get-Epic.ps1
@@ -1,11 +1,19 @@
 function Get-Epic {
     # .ExternalHelp ..\JiraAgilePS-help.xml
-    [CmdletBinding()]
+    [CmdletBinding(SupportsPaging, DefaultParameterSetName = '_ById')]
     [OutputType([AtlassianPS.JiraAgilePS.Epic])]
     param(
-        [Parameter(Position = 0, Mandatory, ValueFromPipeline)]
+        [Parameter(Position = 0, Mandatory, ValueFromPipeline, ParameterSetName = '_ById')]
         [AtlassianPS.JiraAgilePS.Epic[]]
         $Epic,
+
+        [Parameter(Position = 0, Mandatory, ValueFromPipeline, ParameterSetName = '_ByBoard')]
+        [AtlassianPS.JiraAgilePS.Board]
+        $Board,
+
+        [Parameter(ParameterSetName = '_ByBoard')]
+        [ValidateRange(1, 4294967295)]
+        [UInt32]$PageSize = $script:DefaultPageSize,
 
         [Parameter()]
         [System.Management.Automation.PSCredential]
@@ -17,28 +25,58 @@ function Get-Epic {
         Write-Verbose "[$($MyInvocation.MyCommand.Name)] Function started"
 
         $server = Get-JiraConfigServer -ErrorAction Stop
-        $resourceUrl = "$server/rest/agile/1.0/epic/{0}"
+        $resourceUrl_ById = "$server/rest/agile/1.0/epic/{0}"
+        $resourceUrl_ByBoard = "$server/rest/agile/1.0/board/{0}/epic"
     }
 
     process {
         Write-DebugMessage "[$($MyInvocation.MyCommand.Name)] ParameterSetName: $($PsCmdlet.ParameterSetName)"
         Write-DebugMessage "[$($MyInvocation.MyCommand.Name)] PSBoundParameters: $($PSBoundParameters | Out-String)"
 
-        foreach ($_epic in $Epic) {
-            Write-Verbose "[$($MyInvocation.MyCommand.Name)] Processing [$($_epic.Id)]"
-            Write-Debug "[$($MyInvocation.MyCommand.Name)] Processing `$_epic [$($_epic.Id)]"
+        switch ($PSCmdlet.ParameterSetName) {
+            '_ById' {
+                foreach ($_epic in $Epic) {
+                    Write-Verbose "[$($MyInvocation.MyCommand.Name)] Processing [$($_epic.Id)]"
+                    Write-Debug "[$($MyInvocation.MyCommand.Name)] Processing `$_epic [$($_epic.Id)]"
 
-            $requestParameter = @{
-                Uri        = $resourceUrl -f $_epic.Id
-                Method     = "GET"
-                Credential = $Credential
-                Cmdlet     = $PSCmdlet
-                Verbose    = $VerbosePreference
-                Debug      = $DebugPreference
+                    $requestParameter = @{
+                        Uri        = $resourceUrl_ById -f $_epic.Id
+                        Method     = "GET"
+                        Credential = $Credential
+                        Cmdlet     = $PSCmdlet
+                        Verbose    = $VerbosePreference
+                        Debug      = $DebugPreference
+                    }
+
+                    Write-Debug "[$($MyInvocation.MyCommand.Name)] Invoking JiraMethod with `$requestParameter"
+                    Invoke-JiraMethod @requestParameter | ConvertTo-Epic
+                }
             }
+            '_ByBoard' {
+                Write-Verbose "[$($MyInvocation.MyCommand.Name)] Processing Board ID [$($Board.Id)]"
+                Write-Debug "[$($MyInvocation.MyCommand.Name)] Processing `$Board [$($Board.Id)]"
 
-            Write-Debug "[$($MyInvocation.MyCommand.Name)] Invoking JiraMethod with `$requestParameter"
-            Invoke-JiraMethod @requestParameter | ConvertTo-Epic
+                $requestParameter = @{
+                    Uri          = $resourceUrl_ByBoard -f $Board.Id
+                    Method       = "GET"
+                    GetParameter = @{
+                        maxResults = $PageSize
+                    }
+                    Paging       = $true
+                    Credential   = $Credential
+                    Cmdlet       = $PSCmdlet
+                    Verbose      = $VerbosePreference
+                    Debug        = $DebugPreference
+                }
+
+                # Paging
+                ($PSCmdlet.PagingParameters | Get-Member -MemberType Property).Name | ForEach-Object {
+                    $requestParameter[$_] = $PSCmdlet.PagingParameters.$_
+                }
+
+                Write-Debug "[$($MyInvocation.MyCommand.Name)] Invoking JiraMethod with `$requestParameter"
+                Invoke-JiraMethod @requestParameter | Get-AgilePageItem | ConvertTo-Epic
+            }
         }
     }