+++ Tests/Functions/Public/Get-Epic.Unit.Tests.ps1
@@ -25,6 +25,8 @@ InModuleScope JiraAgilePS {
             Context "Parameter Types" {
                 It "has a parameter '<parameter>' of type '<type>'" -TestCases @(
                     @{ parameter = "Epic"; type = [AtlassianPS.JiraAgilePS.Epic[]] }
+                    @{ parameter = "Board"; type = [AtlassianPS.JiraAgilePS.Board] }
+                    @{ parameter = "PageSize"; type = [UInt32] }
                     @{ parameter = "Credential"; type = [System.Management.Automation.PSCredential] }
                 ) {
                     $command | Should -HaveParameter $parameter -Type $type
@@ -66,7 +68,43 @@ InModuleScope JiraAgilePS {
                 $result[0] | Should -BeOfType [AtlassianPS.JiraAgilePS.Epic]
                 (@($result | Select-Object -ExpandProperty Key) -join ",") | Should -Be "EPIC-41,EPIC-42"
             }
+
+            It "requests board epics when Board is supplied" {
+                Mock Invoke-JiraMethod -ModuleName JiraAgilePS {
+                    [pscustomobject]@{
+                        values = @(
+                            [pscustomobject]@{
+                                id   = 31
+                                key  = "EPIC-31"
+                                name = "Epic 31"
+                                done = $false
+                                self = "$jiraServer/rest/agile/1.0/epic/31"
+                            },
+                            [pscustomobject]@{
+                                id   = 32
+                                key  = "EPIC-32"
+                                name = "Epic 32"
+                                done = $true
+                                self = "$jiraServer/rest/agile/1.0/epic/32"
+                            }
+                        )
+                    }
+                }
+                $board = [AtlassianPS.JiraAgilePS.Board]::new(5)
+
+                $result = Get-JiraAgileEpic -Board $board
+
+                Should -Invoke -CommandName Invoke-JiraMethod -ModuleName JiraAgilePS -Exactly -Times 1 -Scope It -ParameterFilter {
+                    $Method -eq "GET" -and
+                    $Uri -eq "$jiraServer/rest/agile/1.0/board/5/epic" -and
+                    $Paging
+                }
+
+                @($result).Count | Should -Be 2
+                $result[0] | Should -BeOfType [AtlassianPS.JiraAgilePS.Epic]
+                $result[0].Name | Should -Be "Epic 31"
+                $result[1].Done | Should -BeTrue
+            }
         }
     }
 }
-